Feature/notifications

Approved-by: Michal Struna <michal.l.struna@gmail.com>

+++ src/Utils/Route.ts
@@ -48,7 +48,8 @@ class Route {
                     }
                 })
                 .catch(error => {
-                    response.sendStatus(error)
+                    console.log(error)
+                    response.status(error.code).send(error)
                 })
         )
     }
@@ -97,20 +98,24 @@ class Route {
     /**
      * Generate routes for all entities.
      * @param model Entity model.
+     * @param access Object of access points. It can be default function (Route.all, ...) or custom request handlers (Route.all(), ...).
      * @returns Route group.
      */
-    public static getRouteGroupForAll(model: IUnspecifiedEntityModel, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_ALL): IRouteGroupForAll {
+    public static getRouteGroupForAll(model: Universis.Model.Unspecified, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_ALL): IRouteGroupForAll {
         const routeGroup: IObject<IRequestHandler> = {}
 
-        if (access.get) {
+        // TODO: Map before and map after.
+        if (access.get && typeof access.get !== 'object') {
             routeGroup.get = access.get(Route.getAllHandler(model))
         }
 
         if (access.post) {
-            routeGroup.post = access.post(Route.addHandler(model), _id => ({ _id }))
+            const mapBefore = 'mapBefore' in access.post ? access.post.mapBefore : item => item
+            const handler = 'access' in access.post ? access.post.access : access.post
+            routeGroup.post = handler(request => mapBefore(request), _id => ({ _id }))
         }
 
-        if (access.delete) {
+        if (access.delete && typeof access.delete !== 'object') {
             routeGroup.delete = access.delete(Route.deleteAllHandler(model), count => ({ count }))
         }
 
@@ -120,30 +125,31 @@ class Route {
     /**
      * Generate routes for one entity.
      * @param model Entity model.
+     * @param access Object of access points. It can be default function (Route.all, ...) or custom request handlers (Route.all(), ...).
      * @returns Route group.
      */
-    public static getRouteGroupForOne(model: IUnspecifiedEntityModel, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_ONE): IRouteGroupForOne {
+    public static getRouteGroupForOne(model: Universis.Model.Unspecified, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_ONE): IRouteGroupForOne {
         const routeGroup: IObject<IRequestHandler> = {}
 
-        if (access.get) {
+        if (access.get && typeof access.get !== 'object') {
             routeGroup.get = access.get(Route.getByIdHandler(model))
         }
 
-        if (access.put) {
-            routeGroup.post = access.put(Route.updateByIdHandler(model), false)
+        if (access.put && typeof access.put !== 'object') {
+            routeGroup.put = access.put(Route.updateByIdHandler(model), false)
         }
 
-        if (access.delete) {
+        if (access.delete && typeof access.delete !== 'object') {
             routeGroup.delete = access.delete(Route.deleteByIdHandler(model), false)
         }
 
         return routeGroup
     }
 
-    public static getRouteGroupForCount(model: IUnspecifiedEntityModel, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_COUNT): IRouteGroupForCount {
+    public static getRouteGroupForCount(model: Universis.Model.Unspecified, access: IRouteGroupAccess = Route.DEFAULT_ROUTE_GROUP_ACCESS_FOR_COUNT): IRouteGroupForCount {
         const routeGroup: IObject<IRequestHandler> = {}
 
-        if (access.get) {
+        if (access.get && typeof access.get !== 'object') {
             routeGroup.get = access.get(Route.getCountHandler(model))
         }
 
@@ -155,14 +161,16 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    public static getAllHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
+    public static getAllHandler = (model: Universis.Model.Unspecified): IRouteAction => (
         ({ query }) => (
-            model.getAll(
+            model.get(
                 Route.getFilterFromQuery(query),
-                query.sort,
-                query.order,
-                parseInt(query.limit),
-                parseInt(query.offset)
+                {
+                    sort: query.sort,
+                    reverse: query.reverse,
+                    limit: parseInt(query.limit),
+                    offset: parseInt(query.offset)
+                }
             ))
     )
 
@@ -171,8 +179,8 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static addHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        ({ body }) => model.add(body)
+    private static addHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        () => model.remove({}) // TODO: Filter?
     )
 
     /**
@@ -180,8 +188,8 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static deleteAllHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        () => model.removeAll()
+    private static deleteAllHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        () => model.remove({}) // TODO: Filter?
     )
 
     /**
@@ -189,8 +197,8 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static getByIdHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        ({ params }) => model.get(params.bodyId)
+    private static getByIdHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        ({ params }) => model.getOne({ _id: params.bodyId })
     )
 
     /**
@@ -198,8 +206,8 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static updateByIdHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        ({ params, body }) => model.update(params.bodyId, body)
+    private static updateByIdHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        ({ params, body }) => model.updateOne({ _id: params.bodyId }, body)
     )
 
     /**
@@ -207,8 +215,8 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static deleteByIdHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        ({ params }) => model.remove(params.bodyId, false) // TODO: Optional force?
+    private static deleteByIdHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        ({ params }) => model.removeOne({ _id: params.bodyId })
     )
 
     /**
@@ -216,24 +224,25 @@ class Route {
      * @param model Model for CRUD operations.
      * @returns Default request handler.
      */
-    private static getCountHandler = (model: IUnspecifiedEntityModel): IRouteAction => (
-        ({ params, body }) => model.getCount() // TODO: Filter?
+    private static getCountHandler = (model: Universis.Model.Unspecified): IRouteAction => (
+        () => model.count({}) // TODO: Filter?
     )
 
-    public static getSwaggerRouteGroupForOne(tags, id, schema) {
-        return {
-            'parameters': [
-                {
-                    'in': 'path',
-                    'name': id,
-                    'required': true,
-                    'schema': {
-                        '$ref': '#/components/schemas/Id'
-                    },
-                    'description': 'Unique identifier of item.'
-                }
-            ],
-            'get': {
+    public static getSwaggerRouteGroupForOne(tags: string[], schema: string, pathParameters: string[] = [], routes: string[] = ['get', 'put', 'delete']) {
+        const result = { parameters: [], get: undefined, put: undefined, delete: undefined }
+
+        result.parameters = pathParameters.map(parameter => ({
+            'in': 'path',
+            'name': parameter,
+            'required': true,
+            'schema': {
+                '$ref': '#/components/schemas/Id'
+            },
+            'description': 'Path identificator.'
+        }))
+
+        if (routes.includes('get')) {
+            result.get = {
                 'tags': tags,
                 'summary': 'Get item by ID.',
                 'description': 'Get item by ID.',
@@ -249,72 +258,123 @@ class Route {
                         }
                     },
                     '404': {
-                        'description': 'Body with ID was not found.'
+                        'description': 'Item with ID was not found.'
                     }
                 }
-            },
-            'put': {
-                'tags': ['Bodies'],
-                'summary': 'Update already existing body.',
-                'description': 'Create new body and return ID of created body.',
+            }
+        }
+
+        if (routes.includes('put')) {
+            result.put = {
+                'tags': tags,
+                'summary': 'Update already existing item.',
+                'description': 'Create new item and return ID of created item.',
                 'requestBody': {
                     'content': {
                         'application/json': {
                             'schema': {
-                                '$ref': '#/components/schemas/NewBody'
-                            },
+                                '$ref': '#/components/schemas/' + schema
+                            }
                         }
                     }
                 },
                 'responses': {
                     '204': {
-                        'description': 'Body was successful updated.'
+                        'description': 'Item was successful updated.'
                     },
                     '400': {
                         'description': 'Invalid values.'
                     },
                     '404': {
-                        'description': 'Body with ID was not found.'
+                        'description': 'Item with ID was not found.'
                     },
                     '409': {
-                        'description': 'Body with this name already exists.'
+                        'description': 'Item with this unique value already exists.'
                     }
                 }
-            },
-            'delete': {
-                'tags': ['Bodies'],
-                'summary': 'Delete body by ID.',
-                'description': 'Delete body by ID.',
-                'parameters': [
-                    {
-                        'in': 'query',
-                        'name': 'force',
-                        'schema': {
-                            'type': 'string',
-                            'example': 'true',
-                            'enum': ['true', 'false']
-                        },
-                        'description': 'Body will be deleted with all its children.'
-                    }
-                ],
+            }
+        }
+
+        if (routes.includes('delete')) {
+            result.delete = {
+                'tags': tags,
+                'summary': 'Delete item by ID.',
+                'description': 'Delete item by ID.',
+                'parameters': [],
                 'responses': {
                     '204': {
-                        'description': 'Body was successful deleted.'
+                        'description': 'Item was successful deleted.'
                     },
                     '400': {
-                        'description': 'Body cannot be deleted, because of existing children.'
+                        'description': 'Item cannot be deleted, because of existing dependencies.'
                     },
                     '404': {
-                        'description': 'Body with ID was not found.'
+                        'description': 'Item with ID was not found.'
+                    }
+                }
+            }
+        }
+
+        return result
+    }
+
+    public static getSwaggerRouteGroupForCount(tags: string[]) {
+        return {
+            'parameters': [
+                {
+                    'in': 'query',
+                    'name': 'filter',
+                    'schema': {
+                        'type': 'object',
+                        'additionalProperties': {
+                            'type': 'string'
+                        }
+                    },
+                    'description': 'Filter items by its any property.'
+                }
+            ],
+            'get': {
+                'tags': tags,
+                'summary': 'Get count of all items.',
+                'description': 'Get count of all items.',
+                'responses': {
+                    '200': {
+                        'description': 'Get items count is successful.',
+                        'content': {
+                            'application/json': {
+                                'schema': {
+                                    'type': 'number'
+                                }
+                            }
+                        }
                     }
                 }
             }
         }
     }
 
-    public static getSwaggerRouteGroupForAll(tags, simpleSchema, newSchema) {
-        return {
-            'get': {
+    /**
+     * Generate code for swagger.
+     * @param tags Tags.
+     * @param simpleSchema
+     * @param newSchema
+     * @param routes
+     */
+    public static getSwaggerRouteGroupForAll(tags: string[], simpleSchema: string, newSchema: string, pathParameters: string[] = [], routes: string[] = ['get', 'post', 'delete']) {
+        const result = { parameters: [], get: undefined, post: undefined, delete: undefined }
+
+        result.parameters = pathParameters.map(parameter => ({
+            'in': 'path',
+            'name': parameter,
+            'required': true,
+            'schema': {
+                '$ref': '#/components/schemas/Id'
+            },
+            'description': 'Path identificator.'
+        }))
+
+        if (routes.includes('get')) {
+            result.get = {
                 'tags': tags,
                 'summary': 'Get all items.',
                 'description': 'Get basic objects of all items.',
@@ -383,8 +443,11 @@ class Route {
                         }
                     }
                 }
-            },
-            'post': {
+            }
+        }
+
+        if (routes.includes('post')) {
+            result.post = {
                 'tags': tags,
                 'summary': 'Create new item.',
                 'description': 'Create new item and return ID of created item.',
@@ -420,14 +483,17 @@ class Route {
                         'description': 'Duplicate values.'
                     }
                 }
-            },
-            'delete': {
+            }
+        }
+
+        if (routes.includes('delete')) {
+            result.delete = {
                 'tags': tags,
                 'summary': 'Delete all items.',
                 'description': 'Delete all items and return count of deleted items.',
                 'responses': {
                     '200': {
-                        'description': 'Items was successful deleted.',
+                        'description': 'Items were successful deleted.',
                         'content': {
                             'application/json': {
                                 'schema': {
@@ -448,6 +514,8 @@ class Route {
                 }
             }
         }
+
+        return result
     }
 
 }
